Update RL implementation and tests
