+++ test/ut/retiarii/test_strategy.py
@@ -58,7 +58,7 @@ def _reset_execution_engine(engine=None):
 
 
 class Net(nn.Module):
-    def __init__(self, hidden_size=32):
+    def __init__(self, hidden_size=32, diff_size=False):
         super(Net, self).__init__()
         self.conv1 = nn.Conv2d(1, 20, 5, 1)
         self.conv2 = nn.Conv2d(20, 50, 5, 1)
@@ -69,7 +69,7 @@ class Net(nn.Module):
         self.fc2 = nn.LayerChoice([
             nn.Linear(hidden_size, 10, bias=False),
             nn.Linear(hidden_size, 10, bias=True)
-        ], label='fc2')
+        ] + ([] if not diff_size else [nn.Linear(hidden_size, 10, bias=False)]), label='fc2')
 
     def forward(self, x):
         x = F.relu(self.conv1(x))
@@ -82,8 +82,8 @@ class Net(nn.Module):
         return F.log_softmax(x, dim=1)
 
 
-def _get_model_and_mutators():
-    base_model = Net()
+def _get_model_and_mutators(**kwargs):
+    base_model = Net(**kwargs)
     script_module = torch.jit.script(base_model)
     base_model_ir = convert_to_graph(script_module, base_model)
     base_model_ir.evaluator = DebugEvaluator()
@@ -139,7 +139,25 @@ def test_evolution():
     _reset_execution_engine()
 
 
+def test_rl():
+    rl = strategy.PolicyBasedRL(max_collect=2, trial_per_collect=10)
+    engine = MockExecutionEngine(failure_prob=0.2)
+    _reset_execution_engine(engine)
+    rl.run(*_get_model_and_mutators(diff_size=True))
+    wait_models(*engine.models)
+    _reset_execution_engine()
+
+    rl = strategy.PolicyBasedRL(max_collect=2, trial_per_collect=10, asynchronous=False)
+    engine = MockExecutionEngine(failure_prob=0.2)
+    _reset_execution_engine(engine)
+    rl.run(*_get_model_and_mutators())
+    wait_models(*engine.models)
+    _reset_execution_engine()
+
+
+
 if __name__ == '__main__':
-    test_grid_search()
-    test_random_search()
-    test_evolution()
+    # test_grid_search()
+    # test_random_search()
+    # test_evolution()
+    test_rl()
